+++ src/grpc_generator/from_proto.rs
@@ -8,7 +8,7 @@ use prost_reflect::prost_types::{
 
 use crate::blueprint::GrpcMethod;
 use crate::config::{Arg, Field, Grpc, Type};
-use crate::generator::proto_generator::{
+use crate::grpc_generator::proto_generator::{
     ConfigWrapper, DescriptorType, Options, ProtoGeneratorConfig, DEFAULT_SPECTATOR,
 };
 
@@ -344,8 +344,10 @@ mod test {
 
     use prost_reflect::prost_types::{FileDescriptorProto, FileDescriptorSet};
 
-    use crate::generator::from_proto::prebuild_config;
-    use crate::generator::proto_generator::{Options, ProtoGeneratorConfig, ProtoGeneratorFxn};
+    use crate::grpc_generator::from_proto::prebuild_config;
+    use crate::grpc_generator::proto_generator::{
+        Options, ProtoGeneratorConfig, ProtoGeneratorFxn,
+    };
 
     fn get_proto_file_descriptor(name: &str) -> anyhow::Result<FileDescriptorProto> {
         let mut proto_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
