+++ src/grpc_generator/snapshots/tailcall__grpc_generator__from_proto__test__from_proto.snap.new
@@ -1,5 +1,6 @@
 ---
-source: src/generator/from_proto.rs
+source: src/grpc_generator/from_proto.rs
+assertion_line: 398
 expression: result
 ---
 schema @server @upstream {
