+++ src/grpc_generator/snapshots/tailcall__grpc_generator__from_proto__test__required_types.snap.new
@@ -1,5 +1,6 @@
 ---
-source: src/generator/from_proto.rs
+source: src/grpc_generator/from_proto.rs
+assertion_line: 469
 expression: cfg
 ---
 schema @server @upstream {
