+++ src/lib.rs
@@ -19,9 +19,9 @@ pub mod data_loader;
 pub mod directive;
 pub mod document;
 pub mod endpoint;
-mod generator;
 pub mod graphql;
 pub mod grpc;
+mod grpc_generator;
 pub mod has_headers;
 pub mod helpers;
 pub mod http;
